+++ Ivan_Zolkin_3823B1PM1/Lab_1_sort/merge_sort/merge_sort.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.6.33829.357
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "merge_sort", "merge_sort.vcxproj", "{8A25635A-C6D8-42E0-ACA8-1C8E2C0B1065}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{8A25635A-C6D8-42E0-ACA8-1C8E2C0B1065}.Debug|x64.ActiveCfg = Debug|x64
+		{8A25635A-C6D8-42E0-ACA8-1C8E2C0B1065}.Debug|x64.Build.0 = Debug|x64
+		{8A25635A-C6D8-42E0-ACA8-1C8E2C0B1065}.Debug|x86.ActiveCfg = Debug|Win32
+		{8A25635A-C6D8-42E0-ACA8-1C8E2C0B1065}.Debug|x86.Build.0 = Debug|Win32
+		{8A25635A-C6D8-42E0-ACA8-1C8E2C0B1065}.Release|x64.ActiveCfg = Release|x64
+		{8A25635A-C6D8-42E0-ACA8-1C8E2C0B1065}.Release|x64.Build.0 = Release|x64
+		{8A25635A-C6D8-42E0-ACA8-1C8E2C0B1065}.Release|x86.ActiveCfg = Release|Win32
+		{8A25635A-C6D8-42E0-ACA8-1C8E2C0B1065}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {994FF38A-E005-4C2B-937B-A99952773947}
+	EndGlobalSection
+EndGlobal
