+++ Ivan_Zolkin_3823B1PM1/Lab_1_sort/selection_sort/selection_sort.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.6.33829.357
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "selection_sort", "selection_sort.vcxproj", "{34E91604-3615-4D6B-AE15-27FF47D8250F}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{34E91604-3615-4D6B-AE15-27FF47D8250F}.Debug|x64.ActiveCfg = Debug|x64
+		{34E91604-3615-4D6B-AE15-27FF47D8250F}.Debug|x64.Build.0 = Debug|x64
+		{34E91604-3615-4D6B-AE15-27FF47D8250F}.Debug|x86.ActiveCfg = Debug|Win32
+		{34E91604-3615-4D6B-AE15-27FF47D8250F}.Debug|x86.Build.0 = Debug|Win32
+		{34E91604-3615-4D6B-AE15-27FF47D8250F}.Release|x64.ActiveCfg = Release|x64
+		{34E91604-3615-4D6B-AE15-27FF47D8250F}.Release|x64.Build.0 = Release|x64
+		{34E91604-3615-4D6B-AE15-27FF47D8250F}.Release|x86.ActiveCfg = Release|Win32
+		{34E91604-3615-4D6B-AE15-27FF47D8250F}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {51188207-72BE-4E9C-89C9-2D30B526863C}
+	EndGlobalSection
+EndGlobal
