+++ Ivan_Zolkin_3823B1PM1/Lab_1_sort/selection_sort/selection_sort.vcxproj.filters
