--- Ivan_Zolkin_3823B1PM1/comb_sort/comb_sort.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.6.33829.357
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "comb_sort", "comb_sort.vcxproj", "{7ADDD666-903E-49D5-BBA5-198C039640FB}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{7ADDD666-903E-49D5-BBA5-198C039640FB}.Debug|x64.ActiveCfg = Debug|x64
-		{7ADDD666-903E-49D5-BBA5-198C039640FB}.Debug|x64.Build.0 = Debug|x64
-		{7ADDD666-903E-49D5-BBA5-198C039640FB}.Debug|x86.ActiveCfg = Debug|Win32
-		{7ADDD666-903E-49D5-BBA5-198C039640FB}.Debug|x86.Build.0 = Debug|Win32
-		{7ADDD666-903E-49D5-BBA5-198C039640FB}.Release|x64.ActiveCfg = Release|x64
-		{7ADDD666-903E-49D5-BBA5-198C039640FB}.Release|x64.Build.0 = Release|x64
-		{7ADDD666-903E-49D5-BBA5-198C039640FB}.Release|x86.ActiveCfg = Release|Win32
-		{7ADDD666-903E-49D5-BBA5-198C039640FB}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {9863B553-EAB2-421D-9570-22CA949D65CE}
-	EndGlobalSection
-EndGlobal
